Refactor OpenAPI plugin code. Fix how schemas are handled for OpenAPI plugins. Remove unused kernel json schema class.
