+++ python/semantic_kernel/connectors/openapi_plugin/models/rest_api_operation_parameter_style.py
@@ -1,0 +1,10 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from enum import Enum
+
+from semantic_kernel.utils.experimental_decorator import experimental_class
+
+
+@experimental_class
+class RestApiOperationParameterStyle(Enum):
+    SIMPLE = "simple"
