+++ python/semantic_kernel/connectors/openapi_plugin/models/rest_api_operation_run_options.py
@@ -1,0 +1,12 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from semantic_kernel.utils.experimental_decorator import experimental_class
+
+
+@experimental_class
+class RestApiOperationRunOptions:
+    """The options for running the REST API operation."""
+
+    def __init__(self, server_url_override=None, api_host_url=None):
+        self.server_url_override: str = server_url_override
+        self.api_host_url: str = api_host_url
