+++ python/semantic_kernel/functions/kernel_function_metadata.py
@@ -10,7 +10,7 @@ from semantic_kernel.utils.validation import FUNCTION_NAME_REGEX, PLUGIN_NAME_RE
 
 
 class KernelFunctionMetadata(KernelBaseModel):
-    name: str = Field(pattern=FUNCTION_NAME_REGEX)
+    name: str = Field(..., pattern=FUNCTION_NAME_REGEX)
     plugin_name: str | None = Field(None, pattern=PLUGIN_NAME_REGEX)
     description: str | None = Field(default=None)
     parameters: list[KernelParameterMetadata] = Field(default_factory=list)
