+++ python/semantic_kernel/functions/kernel_parameter_metadata.py
@@ -23,12 +23,13 @@ class KernelParameterMetadata(KernelBaseModel):
     @classmethod
     def form_schema(cls, data: Any) -> Any:
         if isinstance(data, dict):
-            type_object = data.get("type_object", None)
-            type_ = data.get("type_", None)
-            default_value = data.get("default_value", None)
-            description = data.get("description", None)
-            inferred_schema = cls.infer_schema(type_object, type_, default_value, description)
-            data["schema_data"] = inferred_schema
+            if data.get("schema_data") is None:
+                type_object = data.get("type_object", None)
+                type_ = data.get("type_", None)
+                default_value = data.get("default_value", None)
+                description = data.get("description", None)
+                inferred_schema = cls.infer_schema(type_object, type_, default_value, description)
+                data["schema_data"] = inferred_schema
         return data
 
     @classmethod
