+++ python/semantic_kernel/schema/kernel_json_schema_builder.py
@@ -18,6 +18,7 @@ TYPE_MAPPING = {
     "list": "array",
     "dict": "object",
     "object": "object",
+    "array": "array",
 }
